// part9/Ilari/frontend-flight-diary/src/types.tsx
export enum Weather {
  Sunny = "sunny",
  Rainy = "rainy",
  Cloudy = "cloudy",
  Stormy = "stormy",
  Windy = "windy",
}
export enum Visibility {
  Great = "great",
  Good = "good",
  Ok = "ok",
  Poor = "poor",
}

export interface Diary {
  id: number;
  date: string;
  weather: Weather;
  visibility: Visibility;
  comment?: string;
}
export type NewDiary = Omit<Diary, "id">;

export interface ValidationError {
  message: string;
  errors: Record<string, string[]>;
}